requirements.txt
opencv-python
numpy
matplotlib
pillow
nibabel
scipy
pytest
docopt
